+++ CHANGELOG.md
@@ -20,8 +20,10 @@ HEAD
 * Added the ability to create/assign a `StaticJsonDocument`/`DynamicJsonDocument` from a `JsonArray`/`JsonObject`/`JsonVariant`
 * Added `JsonDocument::isNull()`
 * Added `JsonDocument::operator[]`
+* Added `createNestedArray()` and `createNestedObject()` to `JsonDocument` and `JsonVariant`
 * Added `ARDUINOJSON_TAB` to configure the indentation character
 * Reduced the size of the pretty JSON serializer
+* `JsonVariant` automatically promotes to `JsonObject` or `JsonArray` on write
 
 > ### BREAKING CHANGES
 >